+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### `SAP CX Logging` enhancements
 - Display a log level status from a remote server [1457](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1457)
 - Display an inherited log level status from a remote server [1475](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1475)
+- Reset Loggers cache on connection change [1479](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1479)
 
 ### `items.xml` enhancements
 - Show attribute non-property persistence for contribution results [#1448](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1448)